requirements.txt
requests
beautifulsoup4
tqdm
pillow
azure-storage-blob
spacy
playwright
python-dotenv
azure-storage-blob
imagehash
numpy
openai